requirements.txt
# Core
pyspark>=3.5.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# NLP
nltk>=3.8.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
wordcloud>=1.9.0

# Dashboard
streamlit>=1.28.0

# Scheduling
schedule>=1.2.0

# Testing
pytest>=7.4.0
